Added conditional before clustering or extracting information, to make sure the article has a minimum length.
Long explanation: This check is done because if a user mistypes something in a short phrase and it's not an existing AIML input, then it is considered that the user is sending an article to the system when in reality they were not trying to do that. So to solve this, we check that a string that should be an article has a minimum length.